Shanoir-issue#793 Add protection to study deletion
- Delete from detail and not list
- Disallow deletion is an exam is still present
- re-create study_exam link using rabbitMQ

+++ shanoir-ng-studies/src/main/java/org/shanoir/ng/study/controler/StudyApiController.java
@@ -120,6 +120,11 @@ public class StudyApiController implements StudyApi {
 	@Override
 	public ResponseEntity<Void> deleteStudy(@PathVariable("studyId") Long studyId) {
 		try {
+			Study study = studyService.findById(studyId);
+			if (study.getExaminationIds() != null && !study.getExaminationIds().isEmpty()) {
+				// Error => should not be able to do this see #793
+				return new ResponseEntity<>(HttpStatus.UNPROCESSABLE_ENTITY);
+			}
 			this.deleteProtocolFile(studyId);
 			bidsService.deleteBids(studyId);
 			studyService.deleteById(studyId);
